requirements.txt
python-dotenv
ruff
requests
